refactor(core): split pipeline diagnostics catalogs by severity
Move the large error and warning tables into sibling modules so feature
work that adds a single code no longer edits a 500-line megatable.
Keep advisories and CLI diagnostics on the public diagnostics.ts surface.

Harden the completeness suite: exclude pure catalog modules from the
dead-entry corpus and treat style-${ScaleWarningCode} template emissions
as real emit sites (scale-style-discrete).
